requirements.txt
# ROBIN Benchmark Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.0.0

# Data Processing
datasets>=2.16.0
pandas>=2.0.0
numpy>=1.24.0

# HuggingFace
transformers>=4.36.0
huggingface-hub>=0.20.0
tokenizers>=0.15.0

# Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
evaluate>=0.4.0

# API Clients
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0
together>=0.2.0
groq>=0.4.0

# NLP Tools
regex>=2023.12.0
nltk>=3.8.0
spacy>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Visualization (for Phase 4)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Optional: Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
